pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web-novel-scraper"
dynamic = ["version"]
description = "Python tool that allows you to scrape web novels from various sources and save them to more readable formats like EPUB."
authors = [
    { name="ImagineBrkr", email="salvattore_25@hotmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "Web Novel",
    "Web Novel Scraper",
    "Scraper",
    "Web Novel Downloader",
    "Novel Downloader",
]
dependencies = [
    "python-dotenv",
    "requests",
    "bs4>=0.0.2",
    "ebooklib>=0.18,<1",
    "click>=8.0,<9",
    "platformdirs",
    "dataclasses_json>=0.6.7,<1",
    "ftfy>=6.3.1,<7",
    "Flask>=3.0.0,<4",
    "Flask-SQLAlchemy>=3.1.1,<4",
    "SQLAlchemy>=2.0.23,<3",
    "sentence-transformers>=2.2.2,<3",
    "transformers>=4.35.2,<5",
    "faiss-cpu>=1.8.0,<2",
    "spacy>=3.7.2,<4",
    "textstat>=0.7.3,<1",
    "torch>=2.1.1,<3"
]


[project.scripts]
web-novel-scraper = "web_novel_scraper.__main__:cli"

[project.urls]
Homepage = "https://github.com/ImagineBrkr/web-novel-scraper"
Documentation = "https://web-novel-scraper.readthedocs.io"
Repository = "https://github.com/ImagineBrkr/web-novel-scraper.git"

[tool.hatch.version]
path = "web_novel_scraper/version.py"

[tool.hatch.build.targets.wheel]
packages = ["web_novel_scraper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
